+++ module/data/actor/actor-base.mjs
@@ -13,10 +13,10 @@ export default class HowToBeAHeroActorBase extends AbstractHowToBeAHeroActorData
  * @property {number} attributes.health.temp                  Temporary HP applied on top of value.
  * @property {number} attributes.health.tempmax               Temporary change to the maximum HP.
  * @property {number} attributes.inspiration                  Does this character have inspiration?
- * @property {object} attributes.talents                      
- * @property {number} attributes.talents.knowledge            Contains actors total knowledge value
- * @property {number} attributes.talents.action              Contains actors total action value
- * @property {number} attributes.talents.social               Contains actors total social value
+ * @property {object} attributes.skillSets                      
+ * @property {number} attributes.skillSets.knowledge            Contains actors total knowledge value
+ * @property {number} attributes.skillSets.action              Contains actors total action value
+ * @property {number} attributes.skillSets.social               Contains actors total social value
  * @property {object} details
  * @property {string} details.background                      Character's background.
  * @property {string} details.alignment                       Character's alignment.
@@ -45,21 +45,21 @@ static defineSchema() {
       status: new fields.BooleanField({required: true, label: "HOW_TO_BE_A_HERO.Attributes.InspirationStatus"}),
       value: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.Attributes.Inspiration"})
     }),
-    talents: new fields.SchemaField({
+    skillSets: new fields.SchemaField({
       knowledge: new fields.SchemaField({
-        value: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.Talents.Knowledge.long"}),
-        bonus: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.Talents.Knowledge.bonus"}),
-        totalValue: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.Talents.Knowledge.total"}),
+        value: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.SkillSets.Knowledge.long"}),
+        bonus: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.SkillSets.Knowledge.bonus"}),
+        totalValue: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.SkillSets.Knowledge.total"}),
       }),
       action: new fields.SchemaField({
-        value: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.Talents.Action.long"}),
-        bonus: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.Talents.Action.bonus"}),
-        totalValue: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.Talents.Action.total"}),
+        value: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.SkillSets.Action.long"}),
+        bonus: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.SkillSets.Action.bonus"}),
+        totalValue: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.SkillSets.Action.total"}),
       }),
       social: new fields.SchemaField({
-        value: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.Talents.Social.long"}),
-        bonus: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.Talents.Social.bonus"}),
-        totalValue: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.Talents.Social.total"}),
+        value: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.SkillSets.Social.long"}),
+        bonus: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.SkillSets.Social.bonus"}),
+        totalValue: new fields.NumberField({...requiredInteger, initial: 0, label: "HOW_TO_BE_A_HERO.SkillSets.Social.total"}),
       })
     })
   });
@@ -81,7 +81,7 @@ static defineSchema() {
       name: this.parent.name,
       health: this.attributes.health,
       armor: this.attributes.armor,
-      talents: this.attributes.talents
+      skillSets: this.attributes.skillSets
     });
 
     return {