+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## 0.5.0
+- renamed talents to skill set (#26)
 - update licence to CC BY-NC-SA 4.0 (#19)
 - switch _action_ and _knowledge_ skill set to allign with official character sheet (#25)
 - Removed unlicenced Dungeons and Dragons Artworks (#36)